% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{hoaxy_key}
\alias{hoaxy_key}
\title{Setup}
\usage{
hoaxy_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://rapidapi.com/}.}
}
\description{
Setup your session, all subsequent calls will be done using the API key.
}
\note{
You can specify \code{RAPIDAPI_API_KEY} as environment variable, likely in your \code{.Renviron} file.
}
\examples{
\dontrun{
newsapi_key("xXXxxXxXxXXx")  
}

}
