% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{hx_latest_articles}
\alias{hx_latest_articles}
\title{Top Articles}
\usage{
hx_latest_articles(past_hours = 2, domains = NULL)
}
\arguments{
\item{past_hours}{Hours from "now" (\code{Sys.time()}), you are suggested to start this number from 2 as servers are not real time.}

\item{domains}{Return articles restrained to these domains. Default is None, return all available articles. 
If \code{fact_checking}, return only fact checking articles. If \code{claim}, return only claim articles. 
If \code{fake}, return articles from sites marking as fake by our configuration. You can also input a domain list.}
}
\description{
Return top 20 most active user for the last 30 days.
}
\examples{
\dontrun{
hx_latest_articles()
}

}
