% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{hx_timeline}
\alias{hx_timeline}
\title{Timeline}
\usage{
hx_timeline(ids, resolution = c("D", "M", "W", "H"))
}
\arguments{
\item{ids}{A list or vector of article ids to query, see \code{\link{hx_articles}}.}

\item{resolution}{The resolution of timeline. \code{H}: hour, \code{D}: day, \code{W}: week, \code{M}: month.}
}
\description{
Return timeline of tweets on given articles.
}
\examples{
\dontrun{
articles <- hx_articles("pizzagate")
tl <- hx_timeline(articles$id[1:5])
}

}
