% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{hx_top_articles}
\alias{hx_top_articles}
\title{Top Articles}
\usage{
hx_top_articles(upper_day = NULL, most_recent = FALSE,
  exclude_tags = NULL)
}
\arguments{
\item{upper_day}{When calculating the most active users, we consider a 30 days window. 
The right bound controls the position of the window and it is called \code{upper_day}, 
e.g., if \code{upper_day} is set to \code{2016-12-01}, then the window ranges between 
\code{2016-11-01} and \code{2016-12-01}. Input format is \code{yyyy-mm-dd}, and the 
default value is the date of yesterday. Note that the endpoint does not accept any 
input more recent than the date of yesterday. Also currently the minimal upper_day value is \code{2016-12-12}.}

\item{most_recent}{When set to \code{TRUE}, return most recent available top spreaders, if there is no top spreaders for \code{upper_day}.}

\item{exclude_tags}{Use double quote around excluded tags to make sure the URL encoding works on Mashape. 
This parameter controls the filtering of the returned results by excluded tags. 
The tags could be either in format of a list of strings, e.g., \code{list("fake", "satire")}. 
In the latter case, the first element in the tuple indicates the name of the source that tagged the website. 
Please refer to the Hoaxy FAQ for sources.}
}
\description{
Return top 20 most active user for the last 30 days.
}
\examples{
\dontrun{
hx_top_articles()
}

}
