% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{hx_tweets}
\alias{hx_tweets}
\title{Tweets}
\usage{
hx_tweets(ids)
}
\arguments{
\item{ids}{A list or vector of article ids to query, see \code{\link{hx_articles}}.}
}
\description{
Return tweets on specific fake articles.
}
\examples{
\dontrun{
articles <- hx_articles("pizzagate")
tweets <- hx_tweets(articles$id[1:5])
}

}
